[IMP] *: move the `page` helper fixture to the shared conftest
Use it in `test_limit` instead of direct `requests` calls.
